Crate cirrus_fraud_proof

Subspace fraud proof

This crates provides the feature of generating and verifying the execution proof used in the Subspace fraud proof mechanism. The execution is more fine-grained than the entire block execution, block execution hooks (initialize_block and finalize_block) and any specific extrinsic execution are supported.

Functions

check_execution_proof

Runs the execution using the partial state constructed from the given storage proof and returns the execution result.

prove_execution

Returns a storage proof which can be used to reconstruct a partial state trie to re-run the execution by someone who does not own the whole state.